// src/app/shared/enums/routes.enum.ts
export enum ERoutes {
  dashboard = 'dashboard',
  profile = 'profile',
  users = 'users',
  employees = 'employees',
  opportunities = 'opportunities',
  auth = 'auth',
  login = 'login',
  verification = 'verification',
  signup = 'signup',
  forgotPassword = 'forgot-password',
  resetPassword = 'reset-password',
  anonymous = 'public',
  admin = 'admin',
  investors = 'vendors',
  dvManager = 'dv-manager',
  settings = 'settings',
  holidaysManagement = 'holidays-management',
  plans = 'plans',
  myProfile = 'my-profile',
}